SYS

SYS

SYS(x, ...)

# S3 method for data.frame
SYS(x, group, targetDim, ...)

# S3 method for grouped_df
SYS(x, targetDim, ...)

# S3 method for resample
SYS(x, targetDim, ...)

# S3 method for matrix
SYS(..., group, targetDim, svdMethod = svd,
  covest = Haff_shrinkage)

Arguments

x
data

...
other options

group
grouping variable

targetDim
target dimension to reduce the data to

svdMethod
svd function used for dimesion reduction by default svd in base is used

covest
covariance estimator used forming M matrix

Value

list of reduced data, projection matrix, group variable, discrimination function, m matrix.

Examples

SYS(iris, group = Species, targetDim = 1)

#> $reducedData
#> # A tibble: 150 x 2
#> # Groups:   Species [3]
#>       V1 Species
#>    <dbl> <fctr> 
#>  1 -2.49 setosa 
#>  2 -2.54 setosa 
#>  3 -2.30 setosa 
#>  4 -2.37 setosa 
#>  5 -2.40 setosa 
#>  6 -2.58 setosa 
#>  7 -2.17 setosa 
#>  8 -2.52 setosa 
#>  9 -2.27 setosa 
#> 10 -2.61 setosa 
#> # ... with 140 more rows
#> 
#> $projectionMatrix
#>            [,1]      [,2]       [,3]      [,4]
#> [1,] -0.6083942 0.3358058 -0.4816407 0.5339599
#> 
#> $group
#> Species
#> 
#> $discrimFunc
#> qda
#> 
#> $M
#>                                    Sepal.Length Sepal.Width Petal.Length
#> Sepal.Length -1179.9041 -1641.0719   0.14218367 -0.01403265   0.16654286
#> Sepal.Width    870.6629   745.0807  -0.01403265 -0.04522041   0.07095510
#> Petal.Length -1211.8508 -1095.7056   0.16654286  0.07095510   0.19065714
#> Petal.Width    691.4190  1692.3659   0.04544898  0.03190612   0.06703265
#>              Petal.Width Sepal.Length  Sepal.Width Petal.Length Petal.Width
#> Sepal.Length  0.04544898  0.280093878 -0.005453061   0.28693469  0.03876327
#> Sepal.Width   0.03190612 -0.005453061 -0.039685714   0.05968163  0.03833061
#> Petal.Length  0.06703265  0.286934694  0.059681633   0.27442857  0.04275510
#> Petal.Width   0.02800000  0.038763265  0.038330612   0.04275510  0.06432653
#> 
#> attr(,"class")
#> [1] "reduced"